refactor(radio): move radio-group folder
