+++ src/components.ts
@@ -9,8 +9,7 @@ export { default as Input } from './input';
 export { default as Textarea } from './textarea';
 export { default as Overlay } from './overlay';
 export { default as Progress } from './progress';
-export { default as Radio } from './radio';
-export { default as RadioGroup } from './radio-group';
+export { default as Radio, RadioGroup } from './radio';
 export { default as Rate } from './rate';
 export { default as Switch } from './switch';
 export { default as Stepper } from './stepper';
