--- src/radio-group/radio-group.md
@@ -1,7 +1,0 @@
----
-title: Radio Group 单选按钮组
-# description: 
-spline: base
-isComponent: true
-toc: false
----
