--- src/radio-group/style/index.js
@@ -1,1 +1,0 @@
-import '../../_common/style/mobile/components/radio-group/_index.less';
