+++ src/radio/index.ts
@@ -1,4 +1,5 @@
 import Radio from './radio.vue';
+import _RadioGroup from './radio-group.vue';
 import { withInstall, WithInstallType } from '../shared';
 
 import './style';
@@ -8,6 +9,8 @@ import { TdRadioProps, TdRadioGroupProps } from './type';
 export * from './type';
 export type RadioProps = TdRadioProps;
 export type RadioGroupProps = TdRadioGroupProps;
+
+export const RadioGroup = withInstall(_RadioGroup);
 
 const _Radio: WithInstallType<typeof Radio> = withInstall(Radio);
 export default _Radio;
